requirements.txt
asgiref==3.2.10
astroid==2.4.2
attrs==19.3.0
auditlog3==1.0.1
autopep8==1.5.4
backcall==0.1.0
bleach==3.1.0
boto3==1.15.11
botocore==1.18.11
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.1
cssselect2==0.4.1
decorator==4.4.1
defusedxml==0.6.0
Django==3.0.14
django-admin-sortable2==2.1.3
django-angular==2.3
django-appconf==1.0.3
django-ckeditor==5.8.0
django-classy-tags==2.0.0
django-cms==3.9.0
django-controlcenter==0.2.9
django-crispy-forms==1.7.2
django-debug-toolbar==3.1.1
django-entangled==0.5.3
django-filer==2.2.3
django-filter==2.2.0
django-formtools==2.3
django-fsm==2.8.1
django-fsm-admin==1.2.5
django-hosts==4.0
django-imagekit==4.0.2
django-ipware==4.0.2
django-jet==1.0.8
django-js-asset==1.2.2
django-jsonfield==1.4.0
django-mptt==0.13.4
django-oauth-toolkit==1.3.0
django-pkgconf==0.3.0
django-polymorphic==3.1.0
django-post-office==3.6.1
django-rest-auth==0.9.5
django-sekizai==2.0.0
django-select2==7.10.1
django-shop==1.2.4
django-smartfields==1.1.0
django-storages==1.10.1
django-treebeard==4.5.1
django-widget-tweaks==1.4.8
djangocms-admin-style==2.0.2
djangocms-cascade==1.3.7
djangocms-text-ckeditor==4.0.0
djangorestframework==3.12.1
easy-thumbnails==2.8.3
entrypoints==0.3
facebook-sdk==3.1.0
Faker==4.9.0
html5lib==1.1
idna==2.8
importlib-metadata==0.23
ipykernel==5.1.3
ipython==7.9.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.15.1
Jinja2==2.10.3
jmespath==0.10.0
jsonfield==3.1.0
jsonschema==3.2.0
jupyter-client==5.3.4
jupyter-core==4.6.1
lazy-object-proxy==1.4.3
lxml==4.9.1
Markdown==3.1.1
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
more-itertools==7.2.0
mysqlclient==1.4.6
nb-conda-kernels==2.2.2
nbconvert==5.6.1
nbformat==4.4.0
notebook==6.0.2
oauthlib==3.1.0
pandocfilters==1.4.2
parso==0.5.1
pickleshare==0.7.5
pilkit==2.0
Pillow==6.1.0
prometheus-client==0.7.1
prompt-toolkit==2.0.10
psycopg2==2.8.6
pycodestyle==2.6.0
Pygments==2.4.2
PyJWT==1.7.1
pylint==2.5.3
PyPDF2==1.26.0
pyrsistent==0.15.6
python-dateutil==2.6.0
python-decouple==3.4
python3-openid==3.1.0
pytz==2019.2
pywin32==223
pywinpty==0.5.5
pyzmq==18.1.0
reportlab==3.5.42
requests==2.22.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
Send2Trash==1.5.0
six==1.15.0
social-auth-core==3.2.0
sorl-thumbnail==12.6.3
sqlparse==0.3.1
style==1.1.0
svglib==1.1.0
terminado==0.8.3
testpath==0.4.4
text-unidecode==1.3
tinycss2==1.1.1
toml==0.10.1
tornado==6.0.3
traitlets==4.3.3
typed-ast==1.4.1
Unidecode==1.1.2
update==0.0.1
urllib3==1.25.6
wcwidth==0.1.7
webencodings==0.5.1
wincertstore==0.2
wrapt==1.11.2
xhtml2pdf==0.2.4
zipp==0.6.0
